+++ Engine/Source/Renderer/Renderer.cpp
@@ -59,7 +59,7 @@ namespace Core
             return;
 
         state.screen.BeginFrame();
-        glEnable(GL_DEPTH_TEST);
+
         glEnable(GL_BLEND);
         glBlendFunc(GL_SRC_ALPHA, GL_ONE_MINUS_SRC_ALPHA);
     }