+++ Engine/Source/Renderer/Texture/Texture.h
@@ -25,6 +25,8 @@ namespace Core
         {
             TextureFilter MinFilter;
             TextureFilter MaxFilter;
+
+            Configuration();
         };
 
     private:
@@ -50,7 +52,7 @@ namespace Core
         void Load(const std::string &name, Configuration *config);
         void Load(int width, int height, char channels, CeU8 *data, Configuration *cfg);
 
-        void UpdateWithConfig(Configuration* config);
+        void UpdateWithConfig(Configuration *config);
 
         void Bind();
         void Unbind();
@@ -71,7 +73,7 @@ namespace Core
 
         inline CeU32 GetID() { return id; };
         inline int GetIndex() { return index; };
-        inline Configuration* GetConfig() { return &config; };
+        inline Configuration *GetConfig() { return &config; };
     };
 
 }