+++ Engine/Source/Script/ActorScript.cpp
@@ -1,0 +1,25 @@
+#include "ActorScript.h"
+
+namespace Core
+{
+    ActorScript::ActorScript()
+    {
+        Owner = nullptr;
+    }
+
+    ActorScript::~ActorScript()
+    {
+    }
+
+    void ActorScript::OnStart()
+    {
+    }
+
+    void ActorScript::OnUpdate()
+    {
+    }
+
+    void ActorScript::OnDestroy()
+    {
+    }
+}